+++ src/solvers/dg_structured/dg_3d.jl
@@ -64,9 +64,21 @@ function calc_interface_flux!(u::AbstractArray{<:Any,5}, mesh::StructuredMesh{3}
   @threaded for element in eachelement(dg, cache)
     # Interfaces in negative directions
     # Faster version of "for orientation in (1, 2, 3)"
-    Base.Cartesian.@nexprs 3 orientation->calc_interface_flux!(elements.surface_flux_values, 
-                                                               elements.left_neighbors[orientation, element], 
-                                                               element, orientation, u, mesh, equations, dg)
+
+    # Interfaces in x-direction (`orientation` = 1)
+    calc_interface_flux!(elements.surface_flux_values,
+                         elements.left_neighbors[1, element],
+                         element, 1, u, mesh, equations, dg)
+    
+    # Interfaces in x-direction (`orientation` = 2)
+    calc_interface_flux!(elements.surface_flux_values,
+                         elements.left_neighbors[2, element],
+                         element, 2, u, mesh, equations, dg)
+
+    # Interfaces in x-direction (`orientation` = 3)
+    calc_interface_flux!(elements.surface_flux_values,
+                         elements.left_neighbors[3, element],
+                         element, 3, u, mesh, equations, dg)
   end
 
   return nothing