+++ NEWS.md
@@ -15,7 +15,7 @@ for human readability.
   - multicomponent compressible Euler
   - acoustic perturbation equations
   - Lattice-Boltzmann equations 
-- New mesh type StructuredMesh
+- New structured mesh type `StructuredMesh`
 
 #### Changed